requirements.txt
Flask>=2.3
Flask-Login>=0.6
Flask-SQLAlchemy>=3.1
SQLAlchemy>=2.0
PyMySQL>=1.1
python-dotenv>=1.0
